🔧 refactor: put all colors related states in single composable
