+++ components/uicomponents/ListenNow.vue
@@ -1,7 +1,7 @@
 <script setup>
-import { useColorSelection } from '~/composables/useColorSelection.ts'
+import { useColors } from '~/composables/useColors.ts'
 
-const { accentColor, primaryColor } = useColorSelection()
+const { accentColor, primaryColor } = useColors()
 
 const bg = computed(() =>
   primaryColor.value ? primaryColor.value.hex : '#FFFFFF'
